add first_metadata, and separate metadata init
