+++ crates/polars-stream/src/physical_plan/to_graph.rs
@@ -293,7 +293,7 @@ fn to_graph_rec<'a>(
                     FileScan::Parquet {
                         options,
                         cloud_options,
-                        metadata: _,
+                        metadata: first_metadata,
                     } => {
                         if std::env::var("POLARS_DISABLE_PARQUET_SOURCE").as_deref() != Ok("1") {
                             ctx.graph.add_node(
@@ -305,6 +305,7 @@ fn to_graph_rec<'a>(
                                     options,
                                     cloud_options,
                                     file_options,
+                                    first_metadata.unwrap(),
                                 ),
                                 [],
                             )
